src/terminator/model/TextLine.java: switch to storing style information in a short[] instead of a byte[]...
src/terminator/model/StyledText.java: ...so we have room for two more bits,
controlling whether the foreground or background colors are set. This lets us
distinguish properly between a set color and a default color. In particular,
the colors in "whiptail --yesno food 8 40" are now correct, despite the fact
that our default color scheme is light-on-dark.

src/terminator/model/TextBuffer.java: 
src/terminator/terminal/TerminalListener.java: 
src/terminator/view/highlight/Highlight.java: track byte to short change.

src/terminator/terminal/escape/CSIEscapeAction.java: keep track of whether a
color has actually been set.

+++ terminator/trunk/src/terminator/model/TextBuffer.java
@@ -20,7 +20,7 @@ public class TextBuffer implements TerminalListener {
 	private int width;
 	private int height;
 	private ArrayList textLines = new ArrayList();
-	private int currentStyle = StyledText.getDefaultStyle();
+	private short currentStyle = StyledText.getDefaultStyle();
 	private int firstScrollLineIndex;
 	private int lastScrollLineIndex;
 	private Location caretPosition;
@@ -34,7 +34,7 @@ public class TextBuffer implements TerminalListener {
 	
 	// Fields used for saving and restoring state.
 	private Location savedPosition;
-	private int savedStyle;
+	private short savedStyle;
 	
 	// Fields used for saving and restoring the 'real' screen while the alternative buffer is in use.
 	private TextLine[] savedScreen;
@@ -330,11 +330,11 @@ public class TextBuffer implements TerminalListener {
 		view.setCaretPosition(caretPosition);
 	}
 	
-	public void setStyle(int style) {
+	public void setStyle(short style) {
 		this.currentStyle = style;
 	}
 	
-	public int getStyle() {
+	public short getStyle() {
 		return currentStyle;
 	}
